requirements.txt
fastapi>=0.115,<1.0
openpyxl>=3.1,<4.0
python-multipart>=0.0.9,<1.0
pywebview>=5.4,<6.0
uvicorn>=0.30,<1.0
